requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.2
